+++ src/main/java/baseball/ball/BallResult.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.ball;
 
 public enum BallResult {
     STRIKE, BALL, NOTHING
